Worker/requirements.txt
redis
flask
pyzmq